fix range LBA i32 overflow, range entry with units

+++ sed_manager_config_ui/src/data/locking_range.rs
@@ -15,8 +15,8 @@ impl LockingRange {
         write_locked: bool,
     ) -> Self {
         Self {
-            start_lba: start_lba as i32, // Works up to 1024 TiB. 64-bit would be better.
-            end_lba: end_lba as i32,     // Works up to 1024 TiB. 64-bit would be better.
+            start_lba: start_lba as i64,
+            end_lba: end_lba as i64,
             read_lock_enabled,
             write_lock_enabled,
             read_locked,
